knowledge-base/requirements.txt
# Core dependencies
click>=8.0
pypdf>=3.0.0
gitpython>=3.1.0
anthropic>=0.21.0
python-frontmatter>=1.0.0
requests>=2.31.0
pyyaml>=6.0
magika>=1.0.0

# Optional: PII detection (install with: pip install presidio-analyzer presidio-anonymizer)
# presidio-analyzer>=2.2.0
# presidio-anonymizer>=2.2.0

# Optional: High-quality PDF extraction (highly recommended for PDFs)
# Install with: pip install marker-pdf
# Provides OCR, table extraction, and better formatting
# Falls back to basic pypdf if not available
# marker-pdf>=0.2.0
